LIBSEARCH-152. Moved ItemFormats into WorldCatDiscoveryApiSearcher
Moved the "ItemFormats" class into WorldCatDiscoveryApiSearcher to
fix an issue that was causing the application to hang when running
in the development environment.

The hang appears to have been caused by the Rails autoloading being
unable to find the ItemFormats class when it was in a separate file,
but apparently run forever without returning an error.

https://issues.umd.edu/browse/LIBSEARCH-152