LIBSEARCH-158. Refactored item format algorithm to handle multiple types
Refactor the algorithm for determining the item format to better handle
items with multiple types.

In this implementation, all the fields in the Bib record that may
contain a type (or an array of types) are merged into a single array.
The array is then checked against the "@item_formats" hash, which maps
types to the controlled vocabulary of item formats UMD wants to use.

Because an item may have multiple types (and hence multiple formats),
each format is assigned a weight, with more specific formats having a
higher weight. Where an item has multiple formats, the format with the
highest weight is chosen.

https://issues.umd.edu/browse/LIBSEARCH-158