LIBSEARCH-174. Fix dependencies in "dummy" application for Rails testing
Updated test/dummy/config/application.rb to include "require" files
needed to run the application. Without these changes "rails test"
would fail with "uninitialized constant" errors.

Added "rdf-vocab" as a development dependency because the
"worldcat-discovery" gem has a transitive dependency on it via
the "spira" gem, which only includes it as a development dependency
(and so not automatically included when used with the "dummy"
application).

https://issues.umd.edu/browse/LIBSEARCH-174